clinker-core: extract executor type groups into focused submodules
Pull four loosely-related type groups out of the monolithic
`executor/mod.rs` into dedicated submodules, keeping `mod.rs` as the
re-export root so every `executor::*` path resolves unchanged:

- `transform.rs` — `TransformSpec`, `build_transform_specs`,
  `CompiledTransform`
- `route.rs` — `CompiledRoute`, `CompiledRouteBranch` (fields widened to
  `pub(super)` so the executor's `compile_route` constructor reaches them
  across the new module boundary)
- `dlq.rs` — `DlqEntry`
- `params.rs` — `PipelineRunParams`, `ExecutionReport`, `sum_cpu_io_totals`

Pure relocation: no symbol renames, no public-API change. The moved
`pub` items are re-exported from `mod.rs`, so external callers reaching
`executor::TransformSpec` / `executor::DlqEntry` / `executor::ExecutionReport`
are unaffected, and every existing test passes unmodified.

Closes #258
