+++ crates/clinker-core/src/executor/transform.rs
@@ -1,0 +1,165 @@
+//! Executor-internal transform spec and compiled transform.
+
+use std::sync::Arc;
+
+use crate::config::PipelineConfig;
+use cxl::ast::Statement;
+use cxl::typecheck::TypedProgram;
+
+/// Executor-internal transform spec.
+///
+/// Produced by walking `PipelineConfig::nodes` directly and matching on
+/// `PipelineNode::{Transform, Aggregate, Route}` variants. Used by
+/// executor read paths that still consume a flat Vec-of-transforms view.
+/// The fields are the superset that executor sites read from; see
+/// `build_transform_specs`.
+#[derive(Debug, Clone)]
+pub struct TransformSpec {
+    pub name: String,
+    pub cxl: Option<String>,
+    pub aggregate: Option<crate::config::AggregateConfig>,
+    pub route: Option<crate::config::RouteConfig>,
+    pub input: Option<crate::config::TransformInput>,
+    /// Per-record fan-out ceiling for `emit each` blocks inside this
+    /// transform. Threaded to the evaluator at construction time and
+    /// surfaces as a typed DLQ entry on overflow.
+    pub max_expansion: u64,
+}
+
+impl TransformSpec {
+    pub fn cxl_source(&self) -> &str {
+        if let Some(agg) = &self.aggregate {
+            agg.cxl.as_str()
+        } else if let Some(s) = &self.cxl {
+            s.as_str()
+        } else {
+            ""
+        }
+    }
+}
+
+/// Walk `PipelineConfig::nodes` and materialize a flat `Vec<TransformSpec>`
+/// from the Transform/Aggregate/Route variants in declaration order. Merge
+/// nodes referenced by a transform's input are expanded back into
+/// `TransformInput::Multiple(list)` to match the legacy executor wire shape.
+pub fn build_transform_specs(config: &PipelineConfig) -> Vec<TransformSpec> {
+    use crate::config::node_header::NodeInput;
+    use crate::config::{AggregateConfig, PipelineNode, RouteBranch, RouteConfig, TransformInput};
+
+    let merge_by_name: std::collections::HashMap<&str, Vec<String>> = config
+        .nodes
+        .iter()
+        .filter_map(|n| match &n.value {
+            PipelineNode::Merge { header, .. } => {
+                let upstreams: Vec<String> = header
+                    .inputs
+                    .iter()
+                    .map(|spanned_ni| match &spanned_ni.value {
+                        NodeInput::Single(s) => s.clone(),
+                        NodeInput::Port { node, port } => format!("{node}.{port}"),
+                    })
+                    .collect();
+                Some((header.name.as_str(), upstreams))
+            }
+            _ => None,
+        })
+        .collect();
+
+    let project_input = |ni: &NodeInput| -> Option<TransformInput> {
+        match ni {
+            NodeInput::Single(s) => {
+                if let Some(upstreams) = merge_by_name.get(s.as_str()) {
+                    Some(TransformInput::Multiple(upstreams.clone()))
+                } else {
+                    Some(TransformInput::Single(s.clone()))
+                }
+            }
+            NodeInput::Port { node, port } => {
+                Some(TransformInput::Single(format!("{node}.{port}")))
+            }
+        }
+    };
+
+    let mut out = Vec::new();
+    for spanned in &config.nodes {
+        match &spanned.value {
+            PipelineNode::Transform {
+                header,
+                config: body,
+            } => {
+                out.push(TransformSpec {
+                    name: header.name.clone(),
+                    cxl: Some(body.cxl.as_ref().to_string()),
+                    aggregate: None,
+                    route: None,
+                    input: project_input(&header.input.value),
+                    max_expansion: body.max_expansion,
+                });
+            }
+            PipelineNode::Aggregate {
+                header,
+                config: body,
+            } => {
+                out.push(TransformSpec {
+                    name: header.name.clone(),
+                    cxl: None,
+                    aggregate: Some(AggregateConfig {
+                        group_by: body.group_by.clone(),
+                        cxl: body.cxl.as_ref().to_string(),
+                        strategy: body.strategy,
+                        time_window: body.time_window.clone(),
+                        allowed_lateness: body.allowed_lateness,
+                    }),
+                    route: None,
+                    input: project_input(&header.input.value),
+                    max_expansion: cxl::eval::DEFAULT_MAX_EXPANSION,
+                });
+            }
+            PipelineNode::Route {
+                header,
+                config: body,
+            } => {
+                let branches: Vec<RouteBranch> = body
+                    .conditions
+                    .iter()
+                    .map(|(name, cxl)| RouteBranch {
+                        name: name.clone(),
+                        condition: cxl.as_ref().to_string(),
+                    })
+                    .collect();
+                out.push(TransformSpec {
+                    name: header.name.clone(),
+                    cxl: Some(String::new()),
+                    aggregate: None,
+                    route: Some(RouteConfig {
+                        mode: body.mode,
+                        branches,
+                        default: body.default.clone(),
+                    }),
+                    input: project_input(&header.input.value),
+                    max_expansion: cxl::eval::DEFAULT_MAX_EXPANSION,
+                });
+            }
+            _ => {}
+        }
+    }
+    out
+}
+
+/// Compiled transform: CXL source compiled once, evaluated per record.
+#[derive(Debug)]
+pub struct CompiledTransform {
+    pub(crate) name: String,
+    pub(crate) typed: Arc<TypedProgram>,
+    pub(crate) max_expansion: u64,
+}
+
+impl CompiledTransform {
+    pub(crate) fn has_distinct(&self) -> bool {
+        self.typed
+            .program
+            .statements
+            .iter()
+            .any(|s| matches!(s, Statement::Distinct { .. }))
+    }
+}
